requirements.txt
# Core
opencv-python>=4.8.0
numpy>=1.24.0
pyyaml>=6.0

# Detection & Tracking
ultralytics>=8.1.0
supervision>=0.20.0

# Face Recognition
insightface>=0.7.3
onnxruntime>=1.16.0

# UI & Storage
streamlit>=1.30.0

# Utilities
loguru>=0.7.0
pandas>=2.0.0

# Testing
pytest>=7.0.0
